# dockstore-webservice/src/main/resources/openapi3/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: theglobalalliance@genomicsandhealth.org
    name: Dockstore@ga4gh
    url: https://discuss.dockstore.org/t/opening-helpdesk-tickets/1506
  description: "This describes the dockstore API, a webservice that manages pairs\
    \ of Docker images and associated metadata such as CWL documents and Dockerfiles\
    \ used to build those images. Explore swagger.json for a Swagger 2.0 description\
    \ of our API and explore openapi.yaml for OpenAPI 3.0 descriptions."
  license:
    name: Apache License Version 2.0
    url: https://github.com/dockstore/dockstore/blob/develop/LICENSE
  termsOfService: TBD
  title: Dockstore API
  version: 1.11.0-alpha.2-SNAPSHOT
servers:
- description: Current server when hosted on AWS
  url: /api
  variables: {}
- description: When working locally
  url: /
  variables: {}
- description: Production server
  url: https://dockstore.org/api
  variables: {}
- description: Staging server
  url: https://staging.dockstore.org/api
  variables: {}
- description: Nightly build server
  url: https://dev.dockstore.net/api
  variables: {}
...
paths:
...
  /api/ga4gh/v2/extended/containers/{organization}:
    get:
      description: This endpoint returns entries of an organization.
      operationId: entriesOrgGet
      parameters:
      - in: path
        name: organization
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ToolV1'
          description: An array of Tools of the input organization.
      summary: List entries of an organization
      tags:
      - extendedGA4GH
...
  /api/ga4gh/v2/extended/tools/entry/_search:
    post:
      description: This endpoint searches the indices for all published tools and
        workflows. Used by utilities that expect to talk to an elastic search endpoint.
      operationId: toolsIndexSearch
      requestBody:
        content:
          '*/*':
            schema:
              type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: string
          description: An elastic search result.
      summary: Search the tools and workflows indices.
      tags:
      - extendedGA4GH
  /api/ga4gh/v2/extended/tools/index:
    post:
      description: This endpoint updates the indices for all published tools and workflows.
      operationId: Update the workflows and tools indices
      responses:
        "200":
          content:
            text/plain:
              schema:
                type: integer
                format: int32
          description: Workflows and tools indices populated with entries.
      security:
      - bearer: []
      summary: Update the workflows and tools indices
      tags:
      - extendedGA4GH
  /api/ga4gh/v2/extended/tools/{organization}:
    get:
      description: This endpoint returns tools of an organization.
      operationId: toolsOrgGet
      parameters:
      - in: path
        name: organization
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ToolV1'
          description: An array of Tools of the input organization.
      summary: List tools of an organization
      tags:
      - extendedGA4GH
  /api/ga4gh/v2/extended/workflows/{organization}:
    get:
      description: This endpoint returns workflows of an organization.
      operationId: workflowsOrgGet
      parameters:
      - in: path
        name: organization
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ToolV1'
          description: An array of Tools of the input organization.
      summary: List workflows of an organization
      tags:
      - extendedGA4GH
  /api/ga4gh/v2/extended/{id}/versions/{version_id}/{type}/tests/{relative_path}:
    post:
      description: Test JSON can be annotated with whether they ran correctly keyed
        by platform and associated with some metadata.
      operationId: verifyTestParameterFilePost
      parameters:
      - in: path
        name: type
        required: true
        schema:
          type: string
      - in: path
        name: id
        required: true
        schema:
          type: string
      - in: path
        name: version_id
        required: true
        schema:
          type: string
      - in: path
        name: relative_path
        required: true
        schema:
          type: string
      - in: query
        name: platform
        schema:
          type: string
      - in: query
        name: platform_version
        schema:
          type: string
      - in: query
        name: verified
        schema:
          type: boolean
      - in: query
        name: metadata
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: string
          description: The tool test JSON response.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Credentials not provided or incorrect.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The tool test cannot be found to annotate.
      security:
      - bearer: []
      summary: Annotate test JSON with information on whether it ran successfully
        on particular platforms plus metadata
      tags:
      - extendedGA4GH
...
  /containers/{containerId}:
    delete:
      description: Delete a tool.
      operationId: deleteContainer
      parameters:
      - in: path
        name: containerId
        required: true
        schema:
          type: integer
          format: int64
      responses:
        default:
          content:
            application/json: {}
          description: default response
      security:
      - bearer: []
      tags:
      - containers
    get:
      description: Retrieve a tool
      operationId: getContainer
      parameters:
      - in: path
        name: containerId
        required: true
        schema:
          type: integer
          format: int64
      - in: query
        name: include
        schema:
          type: string
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DockstoreTool'
          description: default response
      security:
      - bearer: []
      tags:
      - containers
    put:
      description: Update the tool with the given tool.
      operationId: updateContainer
      parameters:
      - in: path
        name: containerId
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/DockstoreTool'
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DockstoreTool'
          description: default response
      security:
      - bearer: []
      tags:
      - containers
...
  /organizations:
    get:
      description: "List all organizations that have been approved by a curator or\
        \ admin, sorted by number of stars."
      operationId: getApprovedOrganizations
      responses:
        default:
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Organization'
          description: default response
      summary: List all available organizations.
      tags:
      - organizations
    post:
      description: Create an organization. Organization requires approval by an admin
        before being made public.
      operationId: createOrganization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Organization'
        description: Organization to register.
        required: true
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
          description: default response
      security:
      - bearer: []
      summary: Create an organization.
      tags:
      - organizations
...
  /workflows/{workflowId}:
    get:
      description: Retrieve a workflow
      operationId: getWorkflow
      parameters:
      - in: path
        name: workflowId
        required: true
        schema:
          type: integer
          format: int64
      - in: query
        name: include
        schema:
          type: string
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
          description: default response
      security:
      - bearer: []
      tags:
      - workflows
    put:
      description: Update the workflow with the given workflow.
      operationId: updateWorkflow
      parameters:
      - in: path
        name: workflowId
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/Workflow'
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
          description: default response
      security:
      - bearer: []
      tags:
      - workflows
...
  /workflows/{workflowId}/tools/{workflowVersionId}:
    get:
      description: Get the Tools for a given workflow version.
      operationId: getTableToolContent
      parameters:
      - in: path
        name: workflowId
        required: true
        schema:
          type: integer
          format: int64
      - in: path
        name: workflowVersionId
        required: true
        schema:
          type: integer
          format: int64
      responses:
        default:
          content:
            application/json:
              schema:
                type: string
          description: default response
      security:
      - bearer: []
      tags:
      - workflows
...
components:
  schemas:
    Alias:
      type: object
      properties:
        content:
          type: string
...
    DockstoreTool:
      type: object
      properties:
        aliases:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/Alias'
        author:
          type: string
        checker_id:
          type: integer
          format: int64
        conceptDoi:
          type: string
        custom_docker_registry_path:
          type: string
        customerDockerRegistryPath:
          type: string
          writeOnly: true
        dbCreateDate:
          type: integer
          format: int64
        dbUpdateDate:
          type: integer
          format: int64
        defaultCWLTestParameterFile:
          type: string
        defaultVersion:
          type: string
        defaultWDLTestParameterFile:
          type: string
        default_cwl_path:
          type: string
        default_dockerfile_path:
          type: string
        default_wdl_path:
          type: string
        description:
          type: string
        descriptorType:
          type: array
          items:
            type: string
        email:
          type: string
        forumUrl:
          type: string
          maxLength: 256
          minLength: 0
        gitUrl:
          type: string
        has_checker:
          type: boolean
        id:
          type: integer
          format: int64
        input_file_formats:
          type: array
          items:
            $ref: '#/components/schemas/FileFormat'
          uniqueItems: true
        is_published:
          type: boolean
        labels:
          type: array
          items:
            $ref: '#/components/schemas/Label'
          uniqueItems: true
        lastBuild:
          type: string
          format: date-time
        lastUpdated:
          type: integer
          format: int64
        last_modified:
          type: integer
          format: int32
        last_modified_date:
          type: integer
          format: int64
        licenseInformation:
          $ref: '#/components/schemas/LicenseInformation'
        metadataFromEntry:
          $ref: '#/components/schemas/DockstoreTool'
        metadataFromVersion:
          $ref: '#/components/schemas/Version'
        mode:
          type: string
          enum:
          - AUTO_DETECT_QUAY_TAGS_AUTOMATED_BUILDS
          - AUTO_DETECT_QUAY_TAGS_WITH_MIXED
          - MANUAL_IMAGE_PATH
          - HOSTED
        name:
          type: string
        namespace:
          type: string
        orcidPutCode:
          type: string
        output_file_formats:
          type: array
          items:
            $ref: '#/components/schemas/FileFormat'
          uniqueItems: true
        path:
          type: string
        private_access:
          type: boolean
        registry:
          type: string
          enum:
          - QUAY_IO
          - DOCKER_HUB
          - GITLAB
          - AMAZON_ECR
          - SEVEN_BRIDGES
        registry_string:
          type: string
        starredUsers:
          type: array
          items:
            $ref: '#/components/schemas/User'
          uniqueItems: true
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
          uniqueItems: true
        tool_maintainer_email:
          type: string
        tool_path:
          type: string
        toolname:
          type: string
        topicId:
          type: integer
          format: int64
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
          uniqueItems: true
        workflowVersions:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
          uniqueItems: true
...
    Error:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
      required:
      - code
...
    FileFormat:
      type: object
      properties:
        value:
          type: string
...
    Label:
      type: object
      properties:
        id:
          type: integer
          format: int64
        value:
          type: string
...
    LicenseInformation:
      type: object
      properties:
        licenseName:
          type: string
...
    Organization:
      type: object
      properties:
        aliases:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/Alias'
        avatarUrl:
          type: string
          pattern: "([^\\s]+)(\\.jpg|\\.jpeg|\\.png|\\.gif)"
        collectionsLength:
          type: integer
          format: int64
        dbCreateDate:
          type: integer
          format: int64
        dbUpdateDate:
          type: integer
          format: int64
        description:
          type: string
        displayName:
          type: string
          maxLength: 50
          minLength: 3
          pattern: "[\\w ,_\\-&()']*"
        email:
          type: string
        id:
          type: integer
          format: int64
        link:
          type: string
        location:
          type: string
        name:
          type: string
          maxLength: 39
          minLength: 3
          pattern: "[a-zA-Z][a-zA-Z\\d]*"
        starredUsers:
          type: array
          items:
            $ref: '#/components/schemas/User'
          uniqueItems: true
        status:
          type: string
          enum:
          - PENDING
          - REJECTED
          - APPROVED
        topic:
          type: string
...
    Profile:
      type: object
      properties:
        avatarURL:
          type: string
        bio:
          type: string
        company:
          type: string
        email:
          type: string
        location:
          type: string
        name:
          type: string
        username:
          type: string
...
    ToolClass:
      type: object
      properties:
        description:
          type: string
        id:
          type: string
        name:
          type: string
...
    ToolV1:
      type: object
      properties:
        author:
          type: string
        contains:
          type: array
          items:
            type: string
        description:
          type: string
        id:
          type: string
        meta-version:
          type: string
        organization:
          type: string
        signed:
          type: boolean
        toolclass:
          $ref: '#/components/schemas/ToolClass'
        toolname:
          type: string
        url:
          type: string
        verified:
          type: boolean
        verified-source:
          type: string
        versions:
          type: array
          items:
            $ref: '#/components/schemas/ToolVersionV1'
...
    ToolVersionV1:
      type: object
      properties:
        descriptor-type:
          type: array
          items:
            type: string
            enum:
            - CWL
            - WDL
        dockerfile:
          type: boolean
        id:
          type: string
        image:
          type: string
        meta-version:
          type: string
        name:
          type: string
        url:
          type: string
        verified:
          type: boolean
        verified-source:
          type: string
    User:
      type: object
      properties:
        avatarUrl:
          type: string
        curator:
          type: boolean
        id:
          type: integer
          format: int64
        isAdmin:
          type: boolean
        name:
          type: string
        orcid:
          type: string
        privacyPolicyVersion:
          type: string
          enum:
          - NONE
          - PRIVACY_POLICY_VERSION_2_5
        privacyPolicyVersionAcceptanceDate:
          type: integer
          format: int64
        setupComplete:
          type: boolean
        tosacceptanceDate:
          type: integer
          format: int64
        tosversion:
          type: string
          enum:
          - NONE
          - TOS_VERSION_1
          - TOS_VERSION_2
        tosversionAcceptanceDate:
          type: integer
          format: int64
          writeOnly: true
        userProfiles:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/Profile'
        username:
          type: string
...
  securitySchemes:
    bearer:
      scheme: bearer
      type: http